+++ lib/flowscope.dart
@@ -3,4 +3,5 @@ export 'src/logger/flow_logger.dart';
 export 'src/interceptors/dio_interceptor.dart';
 export 'src/observers/riverpod_observer.dart';
 export 'src/observers/bloc_observer.dart';
+export 'src/observers/route_observer.dart';
 export 'src/core/event.dart';